refactor(observability): split analyze-execution into readers + render modules (refs #3350)
Isolate pure renderers and I/O readers so report formatting is
unit-testable without fixtures. Pure render* functions plus
extractStoryPerfSummaryFromComment now live in
lib/observability/perf-report-render.js; read*/git-scrape helpers live in
lib/observability/perf-report-readers.js. analyze-execution.js is now a
thin CLI wiring runStoryMode/runEpicMode.

extractStoryPerfSummaryFromComment delegates fenced-JSON extraction to
parseFencedJsonComment from structured-comment-parser.js, removing the
open-coded fence regex.
